requirements.txt
fastapi 
pdfplumber
python-dotenv
google-generativeai
base64
